huge refactor in progress
