+++ expander_compiler/Cargo.toml
@@ -54,8 +54,8 @@ name = "trivial_circuit"
 path = "bin/trivial_circuit.rs"
 
 [[bin]]
-name = "expander_serve"
-path = "src/zkcuda/proving_system/expander_parallelized/server_bin.rs"
+name = "expander_serve_vanilla"
+path = "src/zkcuda/proving_system/expander_parallelized/vanilla_bin.rs"
 
 [[bin]]
 name = "matmul_bench"
