+++ expander_compiler/src/zkcuda/proving_system.rs
@@ -10,7 +10,7 @@ pub use dummy::*;
 mod expander;
 pub use expander::*;
 
-pub mod expander_pcs_defered;
+// pub mod expander_pcs_defered;
 
-pub mod expander_parallelized;
-pub use expander_parallelized::*;
+// pub mod expander_parallelized;
+// pub use expander_parallelized::*;
