+++ expander_compiler/src/zkcuda/proving_system/expander_parallelized.rs
@@ -1,7 +1,7 @@
 pub mod client;
 pub mod cmd_utils;
-pub mod server_utils;
 pub mod shared_memory_utils;
+pub mod vanilla_utils;
 
 mod vanilla;
 pub use vanilla::*;
