+++ expander_compiler/src/zkcuda/proving_system/expander_parallelized/client.rs
@@ -1,4 +1,4 @@
-use super::server_utils::{RequestType, SERVER_IP, SERVER_PORT};
+use super::vanilla_utils::{RequestType, SERVER_IP, SERVER_PORT};
 
 use reqwest::Client;
