+++ expander_compiler/src/zkcuda/proving_system/expander_parallelized/vanilla.rs
@@ -10,14 +10,14 @@ use crate::zkcuda::proving_system::expander_parallelized::client::{
 };
 use crate::zkcuda::proving_system::expander_parallelized::cmd_utils::start_server;
 use crate::zkcuda::proving_system::expander_parallelized::shared_memory_utils::SharedMemoryEngine;
-use crate::zkcuda::proving_system::server_utils::get_challenge_for_pcs_with_mpi;
+use crate::zkcuda::proving_system::vanilla_utils::get_challenge_for_pcs_with_mpi;
 use crate::zkcuda::proving_system::{
-    CombinedProof, Commitment, ExpanderGKRCommitment, ExpanderGKRProof, ProvingSystem,
+    CombinedProof, Commitment, ExpanderCommitment, ExpanderProof, ProvingSystem,
 };
 
-use super::super::expander_gkr::{ExpanderGKRProverSetup, ExpanderGKRVerifierSetup};
+use super::super::expander::{ExpanderProverSetup, ExpanderVerifierSetup};
 use super::super::Expander;
-use super::server_utils::{SERVER_IP, SERVER_PORT};
+use super::vanilla_utils::{SERVER_IP, SERVER_PORT};
 use arith::Field;
 use expander_utils::timer::Timer;
 
@@ -31,7 +31,7 @@ use serdes::ExpSerde;
 pub struct ParallelizedExpander<C: GKREngine> {
     _config: std::marker::PhantomData<C>,
 }
-fn parse_port_number() -> u16 {
+pub fn parse_port_number() -> u16 {
     let mut port = SERVER_PORT.lock().unwrap();
     *port = std::env::var("PORT_NUMBER")
         .ok()
@@ -45,10 +45,10 @@ where
     C::FieldConfig: FieldEngine<SimdCircuitField = C::PCSField>,
 {
     fn verify_kernel<ECCConfig: Config<FieldConfig = C::FieldConfig>>(
-        verifier_setup: &ExpanderGKRVerifierSetup<C::PCSField, C::FieldConfig, C::PCSConfig>,
+        verifier_setup: &ExpanderVerifierSetup<C::PCSField, C::FieldConfig, C::PCSConfig>,
         kernel: &Kernel<ECCConfig>,
-        proof: &ExpanderGKRProof,
-        commitments: &[&ExpanderGKRCommitment<C::PCSField, C::FieldConfig, C::PCSConfig>],
+        proof: &ExpanderProof,
+        commitments: &[&ExpanderCommitment<C::PCSField, C::FieldConfig, C::PCSConfig>],
         parallel_count: usize,
         is_broadcast: &[bool],
     ) -> bool {
@@ -112,10 +112,10 @@ where
     fn verify_input_claim<ECCConfig: Config<FieldConfig = C::FieldConfig>>(
         mut proof_reader: impl Read,
         kernel: &Kernel<ECCConfig>,
-        v_keys: &ExpanderGKRVerifierSetup<C::PCSField, C::FieldConfig, C::PCSConfig>,
+        v_keys: &ExpanderVerifierSetup<C::PCSField, C::FieldConfig, C::PCSConfig>,
         challenge: &ExpanderSingleVarChallenge<C::FieldConfig>,
         y: &<C::FieldConfig as FieldEngine>::ChallengeField,
-        commitments: &[&ExpanderGKRCommitment<C::PCSField, C::FieldConfig, C::PCSConfig>],
+        commitments: &[&ExpanderCommitment<C::PCSField, C::FieldConfig, C::PCSConfig>],
         is_broadcast: &[bool],
         parallel_count: usize,
         transcript: &mut C::TranscriptConfig,
@@ -128,7 +128,7 @@ where
             .zip(is_broadcast)
         {
             let val_len =
-                <ExpanderGKRCommitment<C::PCSField, C::FieldConfig, C::PCSConfig> as Commitment<
+                <ExpanderCommitment<C::PCSField, C::FieldConfig, C::PCSConfig> as Commitment<
                     ECCConfig,
                 >>::vals_len(commitment);
             let (challenge_for_pcs, component_idx_vars) =
@@ -191,8 +191,8 @@ impl<C: GKREngine, ECCConfig: Config<FieldConfig = C::FieldConfig>> ProvingSyste
 where
     C::FieldConfig: FieldEngine<SimdCircuitField = C::PCSField>,
 {
-    type ProverSetup = ExpanderGKRProverSetup<C::PCSField, C::FieldConfig, C::PCSConfig>;
-    type VerifierSetup = ExpanderGKRVerifierSetup<C::PCSField, C::FieldConfig, C::PCSConfig>;
+    type ProverSetup = ExpanderProverSetup<C::PCSField, C::FieldConfig, C::PCSConfig>;
+    type VerifierSetup = ExpanderVerifierSetup<C::PCSField, C::FieldConfig, C::PCSConfig>;
     type Proof = CombinedProof<ECCConfig, Expander<C>>;
 
     fn setup(
