+++ expander_compiler/src/zkcuda/proving_system/traits.rs
@@ -16,21 +16,21 @@ pub trait KernelWiseProvingSystem<C: Config> {
     type VerifierSetup: Clone + Send + Sync + ExpSerde;
     type Proof: Clone + Send + Sync + ExpSerde;
     type Commitment: Commitment<C> + Send + Sync + ExpSerde;
-    type CommitmentExtraInfo: Clone + Send + Sync + ExpSerde;
+    type CommitmentState: Clone + Send + Sync + ExpSerde;
 
     fn setup(computation_graph: &ComputationGraph<C>) -> (Self::ProverSetup, Self::VerifierSetup);
 
     fn commit(
         prover_setup: &Self::ProverSetup,
         vals: &[SIMDField<C>],
-    ) -> (Self::Commitment, Self::CommitmentExtraInfo);
+    ) -> (Self::Commitment, Self::CommitmentState);
 
     #[allow(clippy::too_many_arguments)]
     fn prove_kernel(
         prover_setup: &Self::ProverSetup,
         kernel: &Kernel<C>,
         commitments: &[&Self::Commitment],
-        commitments_extra_info: &[&Self::CommitmentExtraInfo],
+        commitments_extra_info: &[&Self::CommitmentState],
         commitments_values: &[&[SIMDField<C>]],
         parallel_count: usize,
         is_broadcast: &[bool],
